+++ CHANGELOG.md
@@ -9,6 +9,8 @@ This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.htm
 - Added the `must_use` attribute to all pure functions.
 - Added the [`num-traits`](https://crates.io/crates/num-traits) and
  [`num-complex`](https://crates.io/crates/num-complex) crates to dependencies.
+- Deprecated the `LambertW` trait, as this crate is not really the place for
+ such API decisions to be made on behalf of others.
  It is also unclear how the trait should be defined given the newly introduced
  general implementation.